+++ src/pages/MyPage/components/Mailbox/Letters/index.tsx
